Modify Security issue

+++ admin/includess/view_all_users.php
@@ -105,11 +105,14 @@ if(isset($_GET['change_to_subscriber'])){
 
 <?php
 
-if(isset($_GET['delete'])){
-    $the_user_id = $_GET['delete'];
-    $query = "DELETE FROM users WHERE user_id ={$the_user_id}";
-    $delete_user = mysqli_query($connection, $query);
-    header("Location: users.php");
-}
-
-//?>
+    if (isset($_GET['delete'])) {
+        if (isset($_SESSION['user_role'])) {
+            if ($_SESSION['user_role'] == 'admin') {
+                $the_user_id = $_GET['delete'];
+                $query = "DELETE FROM users WHERE user_id ={$the_user_id}";
+                $delete_user = mysqli_query($connection, $query);
+                header("Location: users.php");
+            }
+        }
+    }
+    ?>
